+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleToolbar.razor.cs
@@ -6,4 +6,10 @@ public partial class NimbleToolbar
 {
     [Parameter]
     public RenderFragment? ChildContent { get; set; }
+
+    /// <summary>
+    /// Gets or sets a collection of additional attributes that will be applied to the created element.
+    /// </summary>
+    [Parameter(CaptureUnmatchedValues = true)]
+    public IReadOnlyDictionary<string, object>? AdditionalAttributes { get; set; }
 }