+++ packages/nimble-blazor/NimbleBlazor/Direction.cs
@@ -8,7 +8,7 @@ public enum Direction
 
 internal static class DirectionExtensions
 {
-    private static readonly Dictionary<Direction, string> _directionValues = AttributeHelpers.GetEnumNamesAsKebabCaseValues<Direction>();
+    private static readonly Dictionary<Direction, string> _enumValues = AttributeHelpers.GetEnumNamesAsKebabCaseValues<Direction>();
 
-    public static string? ToAttributeValue(this Direction? value) => value == null ? null : _directionValues[value.Value];
+    public static string? ToAttributeValue(this Direction? value) => value == null ? null : _enumValues[value.Value];
 }