+++ packages/nimble-components/src/patterns/error/types.ts
@@ -1,9 +1,14 @@
 /**
- * This interface should be used by components that want to leverage the errorTextTemplate defined in templates.ts
+ * This interface should be used by components that want to leverage the error pattern. The error pattern conditionally adds error text to the template when an error is configured to be visible.
  */
-export interface IHasErrorText {
+export interface ErrorPattern {
     /**
-     * The error text that will be displayed when a component has the 'invalid' class set
+     * The error text that will be displayed when a component is in the error appearance
      */
-    errorText: string | undefined;
+    errorText?: string;
+
+    /*
+     * Show the error appearance of the control
+     */
+    errorVisible: boolean;
 }