+++ packages/nimble-components/src/text-field/styles.ts
@@ -76,7 +76,7 @@ export const styles = css`
         border-color: rgba(${borderRgbPartialColor}, 0.1);
     }
 
-    :host(.invalid) .root {
+    :host([error-visible]) .root {
         border-bottom-color: ${failColor};
     }
 
@@ -84,7 +84,7 @@ export const styles = css`
         border-bottom-color: ${borderHoverColor};
     }
 
-    :host([appearance='frameless'].clear-inline-padding) .root {
+    :host([appearance='frameless'][full-bleed]) .root {
         padding-left: 0px;
         padding-right: 0px;
     }
@@ -97,7 +97,7 @@ export const styles = css`
         user-select: none;
     }
 
-    :host([appearance='frameless'].clear-inline-padding) .root::before {
+    :host([appearance='frameless'][full-bleed]) .root::before {
         display: none;
     }
 
@@ -109,7 +109,7 @@ export const styles = css`
         user-select: none;
     }
 
-    :host([appearance='frameless'].clear-inline-padding) .root::after {
+    :host([appearance='frameless'][full-bleed]) .root::after {
         display: none;
     }
 
@@ -180,7 +180,7 @@ export const styles = css`
         }
     }
 
-    :host(.invalid) [part='end']::after {
+    :host([error-visible]) [part='end']::after {
         border-bottom-color: ${failColor};
     }
 
@@ -223,7 +223,7 @@ export const styles = css`
             }
 
             .root:focus-within,
-            :host(.invalid) .root {
+            :host([error-visible]) .root {
                 border-bottom-width: ${borderWidth};
                 padding-bottom: 0;
             }