+++ packages/nimble-components/src/tooltip/types.ts
@@ -2,10 +2,11 @@
  * Types of tooltip appearance.
  * @public
  */
-export const TooltipStatus = {
-    default: '',
-    fail: 'fail',
+export const TooltipSeverity = {
+    default: undefined,
+    error: 'error',
     information: 'information'
 } as const;
 
-export type TooltipStatus = typeof TooltipStatus[keyof typeof TooltipStatus];
+export type TooltipSeverity =
+    typeof TooltipSeverity[keyof typeof TooltipSeverity];